+++ iOS/G3MiOSSDK/Commons/Basic/IImageUtils.hpp
@@ -56,7 +56,7 @@ public:
                       const std::vector<const IImage*>& images,
                       const std::vector<RectangleF*>& sourceRects,
                       const std::vector<RectangleF*>& destRects,
-                      const std::vector<double>& transparencies,
+                      const std::vector<float>& transparencies,
                       IImageListener* listener,
                       bool autodelete);
 
@@ -64,7 +64,7 @@ public:
                       const std::vector<const IImage*>& images,
                       const std::vector<RectangleF*>& sourceRects,
                       const std::vector<RectangleF*>& destRects,
-                      const std::vector<double>& transparencies,
+                      const std::vector<float>& transparencies,
                       IImageListener* listener,
                       bool autodelete) {
     combine(extent._x, extent._y,
